requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiopg==1.3.3
aioschedule==0.5.2
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
loguru==0.6.0
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2022.1
SQLAlchemy==1.4.35
yarl==1.7.2
